refactor: share claim upload action plumbing
