+++ apps/web/src/features/member/claims/actions.ts
@@ -1,45 +1,16 @@
 'use server';
 
-import {
-  emitClaimAiRunRequestedService,
-  markClaimAiRunDispatchFailedService,
-} from '@/lib/ai/claim-workflows';
 import { auth } from '@/lib/auth';
+import {
+  createSignedUploadUrl,
+  persistClaimDocumentAndQueueWorkflows,
+  revalidatePathForAllLocales,
+} from '@/features/claims/upload/server/shared-upload';
 import { resolveEvidenceBucketName } from '@/lib/storage/evidence-bucket';
-import { LOCALES } from '@/i18n/locales';
-import { claimDocuments, claims, db } from '@interdomestik/database';
-import { queueClaimDocumentAiWorkflows } from '@interdomestik/domain-claims/claims/ai-workflows';
+import { claims, db } from '@interdomestik/database';
 import { ensureTenantId } from '@interdomestik/shared-auth';
-import { createClient } from '@supabase/supabase-js';
-import { randomUUID } from 'crypto';
 import { and, eq } from 'drizzle-orm';
-import { revalidatePath } from 'next/cache';
 import { headers } from 'next/headers';
-
-const SIGNED_UPLOAD_MAX_ATTEMPTS = 3;
-const SIGNED_UPLOAD_RETRY_DELAY_MS = process.env.NODE_ENV === 'test' ? 0 : 250;
-const TRANSIENT_UPLOAD_ERROR_PATTERNS = [
-  /fetch failed/i,
-  /network/i,
-  /timed?\s*out/i,
-  /econnreset/i,
-  /ehostunreach/i,
-  /enotfound/i,
-  /socket hang up/i,
-  /temporar/i,
-  /service unavailable/i,
-  /too many requests/i,
-];
-
-function shouldRetrySignedUpload(message: string): boolean {
-  return TRANSIENT_UPLOAD_ERROR_PATTERNS.some(pattern => pattern.test(message));
-}
-
-async function waitForSignedUploadRetry(attempt: number): Promise<void> {
-  if (SIGNED_UPLOAD_RETRY_DELAY_MS <= 0) return;
-  const delay = SIGNED_UPLOAD_RETRY_DELAY_MS * attempt;
-  await new Promise(resolve => setTimeout(resolve, delay));
-}
 
 export type GenerateUploadUrlResult =
   | { success: true; url: string; path: string; id: string; token: string; bucket: string }
@@ -48,7 +19,7 @@ export type GenerateUploadUrlResult =
 export async function generateUploadUrl(
   claimId: string,
   fileName: string,
-  contentType: string,
+  _contentType: string,
   fileSize: number
 ): Promise<GenerateUploadUrlResult> {
   const session = await auth.api.getSession({
@@ -86,75 +57,14 @@ export async function generateUploadUrl(
     return { success: false, error: 'Claim not found', status: 404 };
   }
 
-  // Validation
-  if (fileSize > 50 * 1024 * 1024) {
-    return { success: false, error: 'File too large (max 50MB)', status: 413 };
-  }
-
-  // Path: pii/tenants/{tenantId}/claims/{claimId}/{uuid}.{ext}
-  const ext = fileName.split('.').pop() || 'bin';
-  const fileId = randomUUID();
-  const path = `pii/tenants/${tenantId}/claims/${claimId}/${fileId}.${ext}`;
-
-  // Supabase
-  const supabaseUrl = process.env.SUPABASE_URL || process.env.NEXT_PUBLIC_SUPABASE_URL;
-  const supabaseServiceKey = process.env.SUPABASE_SERVICE_ROLE_KEY;
-
-  if (!supabaseUrl || !supabaseServiceKey) {
-    console.error('Missing Supabase configuration');
-    return { success: false, error: 'Configuration error', status: 500 };
-  }
-
-  const supabase = createClient(supabaseUrl, supabaseServiceKey);
-
-  try {
-    for (let attempt = 1; attempt <= SIGNED_UPLOAD_MAX_ATTEMPTS; attempt += 1) {
-      const { data, error } = await supabase.storage
-        .from(evidenceBucket)
-        .createSignedUploadUrl(path, { upsert: true });
-
-      if (!error && data?.signedUrl && data?.token) {
-        return {
-          success: true,
-          url: data.signedUrl,
-          path: path,
-          id: fileId,
-          token: data.token,
-          bucket: evidenceBucket,
-        };
-      }
-
-      const detail = error?.message ?? 'Unknown storage error';
-      const retryable = shouldRetrySignedUpload(detail);
-      const hasAttemptsLeft = attempt < SIGNED_UPLOAD_MAX_ATTEMPTS;
-
-      console.error('Supabase signed URL error:', {
-        bucket: evidenceBucket,
-        path,
-        detail,
-        error,
-        attempt,
-        maxAttempts: SIGNED_UPLOAD_MAX_ATTEMPTS,
-        retryable,
-      });
-
-      if (retryable && hasAttemptsLeft) {
-        await waitForSignedUploadRetry(attempt);
-        continue;
-      }
-
-      return { success: false, error: `Failed to generate upload URL: ${detail}`, status: 500 };
-    }
-
-    return {
-      success: false,
-      error: 'Failed to generate upload URL: Unknown storage error',
-      status: 500,
-    };
-  } catch (err) {
-    console.error('generateUploadUrl error:', err);
-    return { success: false, error: 'Unexpected error', status: 500 };
-  }
+  return createSignedUploadUrl({
+    bucket: evidenceBucket,
+    claimId,
+    fileName,
+    fileSize,
+    logPrefix: '[member/claims]',
+    tenantId,
+  });
 }
 
 export type ConfirmUploadResult =
@@ -208,77 +118,6 @@ async function resolveConfirmUploadContext(): Promise<ConfirmUploadContext> {
   }
 }
 
-async function enqueueDocumentAiWorkflowsBestEffort(input: {
-  claimId: string;
-  fileId: string;
-  originalName: string;
-  storagePath: string;
-  mimeType: string;
-  fileSize: number;
-  resolvedBucket: string;
-  category: 'evidence' | 'legal';
-  tenantId: string;
-  userId: string;
-}): Promise<void> {
-  const {
-    claimId,
-    fileId,
-    originalName,
-    storagePath,
-    mimeType,
-    fileSize,
-    resolvedBucket,
-    category,
-    tenantId,
-    userId,
-  } = input;
-
-  try {
-    const queuedRuns = await db.transaction(async tx =>
-      queueClaimDocumentAiWorkflows({
-        tx,
-        claimId,
-        tenantId,
-        userId,
-        files: [
-          {
-            documentId: fileId,
-            name: originalName,
-            path: storagePath,
-            type: mimeType,
-            size: fileSize,
-            bucket: resolvedBucket,
-            category,
-          },
-        ],
-      })
-    );
-
-    for (const queuedRun of queuedRuns) {
-      try {
-        await emitClaimAiRunRequestedService(queuedRun);
-      } catch (error) {
-        await markClaimAiRunDispatchFailedService({
-          runId: queuedRun.runId,
-          message: error instanceof Error ? error.message : 'Failed to dispatch claim AI run.',
-        });
-      }
-    }
-  } catch (queueError) {
-    console.error('[member/claims] confirmUpload AI queue failed after metadata persisted', {
-      claimId,
-      fileId,
-      message: queueError instanceof Error ? queueError.message : String(queueError),
-    });
-  }
-}
-
-function revalidatePathForAllLocales(path: string) {
-  for (const locale of LOCALES) {
-    revalidatePath(`/${locale}${path}`);
-  }
-}
-
 export async function confirmUpload({
   claimId,
   storagePath,
@@ -321,29 +160,16 @@ export async function confirmUpload({
       };
     }
 
-    await db.transaction(async tx => {
-      await tx.insert(claimDocuments).values({
-        id: fileId,
-        tenantId: tenantId,
-        claimId,
-        name: originalName,
-        filePath: storagePath,
-        fileType: mimeType,
-        fileSize,
-        bucket: resolvedBucket,
-        category,
-        uploadedBy: session.user.id,
-      });
-    });
-    await enqueueDocumentAiWorkflowsBestEffort({
+    await persistClaimDocumentAndQueueWorkflows({
+      category,
       claimId,
       fileId,
-      originalName,
-      storagePath,
-      mimeType,
       fileSize,
+      logPrefix: '[member/claims] confirmUpload',
+      mimeType,
+      originalName,
       resolvedBucket,
-      category,
+      storagePath,
       tenantId,
       userId: session.user.id,
     });
